[DEV-2917] Create 2 hooks
